UPSTREAM: tytso: ext4 crypto: check for too-short encrypted file names
An encrypted file name should never be shorter than an 16 bytes, the
AES block size.  The 3.10 crypto layer will oops and crash the kernel
if ciphertext shorter than the block size is passed to it.

Fortunately, in modern kernels the crypto layer will not crash the
kernel in this scenario, but nevertheless, it represents a corrupted
directory, and we should detect it and mark the file system as
corrupted so that e2fsck can fix this.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
(cherry picked from commit 27977b69e493c9b259eb0490534e0f74bc325ba8)

BUG=chromium:347322
TEST=Compiled, Installed on samus
TEST=Test with e2fsprogs 1.43-WIP

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: I36bddab916adf1a5f4ddb6829046821db807ac65
Reviewed-on: https://chromium-review.googlesource.com/321712
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>