UPSTREAM: drm/tegra: sor: Hook up I2C-over-AUX to output
This is necessary for the output abstraction to retrieve a list of valid
modes from the EDID of a connected panel/monitor. This will be useful in
conjunction with DisplayPort support that will be added in a subsequent
patch, so that the driver can read EDID via the AUX channel.

Signed-off-by: Thierry Reding <treding@nvidia.com>
[from https://github.com/thierryreding/linux.git]
(cherry picked from commit d5198802278ee7b80ac3ea4c15b80b6c0ec1d1dc)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

Change-Id: Ib42a22ab130ca7ed93b8c3e1a1f04f3ac00502e6
Reviewed-on: https://chromium-review.googlesource.com/308894
Commit-Ready: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>