CHROMIUM: iwl7000: Allow configuring FW dbg data with an external file
Currently configuring the fw debug mechanism is done
by appending binary data to the ucode file.
This limits customers / users who do not want to change the
first compiled ucode.

In addition to the current way of configuring the collection, enable
doing it with an external file that is separated from the ucode.

Add a dbg-cfg entry which gets as input the name of the file
with the fw debug data configuration, and then parse the data with
the existing fw parser.

Signed-off-by: Golan Ben-Ami <golan.ben.ami@intel.com>
iwl7000-tree: 80fa8eab05e5e2eff6bb5051b16b96024a7911cf