UPSTREAM: ath10k: fix beamformee VHT STS capability
The VHT STS CAP shall be reported by firmware to host, like in case of
QCA99x0. For QCA6174 hw family this isn't set for some reason.
So for this particular chips, let's assume it has the ability to
support VHT NDP in up to 4 STSs (which is true by the way).

Change the published beamformee STS cap accordingly to 3 or to what
the firmware reports.

Assumption so far, it suppose to be the num_rf_chains-1, was
completely wrong.

(cherry picked from commit 707a0c811ca02861f1500f0696efebcf3bd2e2e2)

BUG=chrome-os-partner:49260
TEST=built ath10k modules for x86

Change-Id: I42b2207766f39e05f72c03284ad1f447aba16c97
Signed-off-by: Bartosz Markowski <bartosz.markowski@tieto.com>
Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>
Signed-off-by: Anilkumar Kolli <akolli@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/333490
Commit-Ready: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>