CHROMIUM: fs/pstore: Use memcpy_fromio() to save old ramoops buffer
The ramoops buffer may be mapped as either I/O memory or uncached
memory.  On ARM64, this results in a device-type (strongly-ordered)
mapping.  Since unnaligned accesses to device-type memory will
generate an alignment fault (regardless of whether or not strict
alignment checking is enabled), it is not safe to use memcpy().
memcpy_fromio() is guaranteed to only use aligned accesses, so use
that instead.

BUG=chrome-os-partner:44342
TEST=Manually insert an unaligned copy from the ramoops buffer and
observe that it no longer generates an alignment fault.

Change-Id: Ida6cb9d693f1a555f20e16fce06c967852ed6c68
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/297216
Commit-Ready: Puneet Kumar <puneetster@chromium.org>
Reviewed-by: Puneet Kumar <puneetster@chromium.org>