CHROMIUM: iio: core: add double tap gesture
Add double tap as possible gesture. This is an inteface change: however,
the sysfs entry is based on string, so if other gestures are added on
the trunk in the meantime, we will still be able to merge this change.

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
(cherry picked from commit f378432e0e419e76c1908ba795511d2a4c4231c5)
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>