iommu/io-pgtable-arm: set T1SZ to a valid value in spite of EPD1
Some SMMU implementations perform range checks on T1SZ even if EPD1 is
set, indicating that page-table walks via TTBR1 are disabled.

This patch ensures that we construct a TCR value with the T1SZ field set
to the same value as T0SZ for stage-1 page-tables. On IOMMUs without the
incorrect checking, the setting is harmless.

Cc: <stable@vger.kernel.org>
Reported-by: Matt Evans <matt.evans@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>