UPSTREAM: ath10k: fix peer assoc complete WMI command for 10.4
There is an extra 4-byte member when compared to WMI 10.2 added to
assoc complete command in WMI 10.4. This new member is used for 160Mhz
related configuration. This WMI command mismatch between host and
firmware does not cause any real issues because this new member is not
used in 10.4 firmwares so far (10.4.1.00030-1). This difference in WMI
command interface brings in a new wmi_ops for 10.4 gen_peer_assoc().
No noticeable functionality differences with this change can be seen
with the current 10.4 firmwares, but the WMI interface has to be
fixed to work with future 10.4 firmwares which may be using this new
member.

(cherry picked from commit b54e16f1040adcd7341ab07ca07ee8577144e4fa)

BUG=chrome-os-partner:49260
TEST=built ath10k modules for x86

Change-Id: I5839a7c58c14f4ea0794b775f7bf9c144f982451
Signed-off-by: Vasanthakumar Thiagarajan <vthiagar@qti.qualcomm.com>
Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>
Signed-off-by: Maharaja Kennadyrajan <mkenna@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/333556
Commit-Ready: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>