CHROMIUM: wakeup: Clear wakeup_type in userspace
This allows userspace to clear /sys/power/wakeup_type instead of doing it
in the kernel. If an automatic wakeup cancels suspend, we don't want to
fully resume the system. When clearing it in the kernel, we wouldn't be
able to tell this happened, though. This is because the wakeup_type could
be user from a wakeup before suspend was initiated. It would not change
from an automatic wakeup, so userspace would not know about the automatic
wakeup.

BUG=chrome-os-partner:47095
TEST=on glados:
echo 1 > /usr/share/power_manager/suspend_to_idle
restart powerd
echo automatic > /sys/class/rtc/rtc0/device/power/wakeup_type
echo +10 > /sys/class/rtc/rtc0/wakealarm; echo unknown > \
/sys/power/wakeup_type; powerd_dbus_suspend
-check that /sys/power/wakeup_type is automatic after the system wakes from
the rtc alarm.

Change-Id: Id4a83f70a2ac1fa66e6f5eaeb5e368cd3beac688
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/311392
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>

Conflicts:
	include/linux/suspend.h

[rebase44(groeck): Resolved conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>