CHROMIUM: ASoC: core: Adds support for cpu loopback dai_link
soc_link_dai_widgets() creates codec-codec loopback link which
allows DPCM to trigger these links and program the parameters for
these links These links are:

	Codec1 Tx (Pb)  ----------> Rx Codec2 (Cap)
	Codec2 Tx (Cap) <---------- Tx Codec1 (Pb)

But as we try to model a CPU based loops for non host links like
modem to codec loop, the above mapping gets wrong as:

	CPU Tx (Pb)  -----------> Codec Tx (Pb)
	CPU Rx (Cap) <----------- Codec Rx (Cap)

So we add a new flag to handle above scenario and link widgets as
depecited above

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
(am from https://patchwork.kernel.org/patch/7585761/)

This commit is Nack upstream. but this is currently needed
for ADI spro feature enablement.
Once we have new version of patches, this will be reverted.

BUG=chrome-os-partner:49191
TEST=with the series applied, SSPO loopback is enabled
     to communicate with algorithm in dsp.

Change-Id: I659692ef310a99e1024a2df186d7bb5adc4069fd
Signed-off-by: Sathya Prakash M R <sathya.prakash.m.r@intel.com>
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/323021