CHROMIUM: drm/nouveau/gr: consider ELPG when programming ZBC
When exiting ELPG, it's PMU's responsibility to restore the ZBC
registers for engine. Therefore we have to notify PMU once the registers
are updated, and then PMU can save the latest copy for EPLG exit. The
ELPG is also needed to be disabled before we program ZBC to avoid race.

BUG=chrome-os-partner:47746
TEST=verified zbc color by NVIDIA tool

Change-Id: Idad37cfa7e8c62da64d12a75c57b9cfcabdfd3d4
Signed-off-by: Vince Hsu <vince.h@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/312680
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>