UPSTREAM: regulator: core: fix regulator_lock_supply regression
As noticed by Geert Uytterhoeven, my patch to avoid a harmless build warning
in regulator_lock_supply() was total crap and introduced a real bug:

> [ BUG: bad unlock balance detected! ]
> kworker/u4:0/6 is trying to release lock (&rdev->mutex) at:
> [<c0247b84>] regulator_set_voltage+0x38/0x50

we still lock the regulator supplies, but not the actual regulators,
so we are missing a lock, and the unlock is unbalanced.

This rectifies it by first locking the regulator device itself before
using the same loop as before to lock its supplies.

Reported-by: Geert Uytterhoeven <geert@linux-m68k.org>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Fixes: 716fec9d1965 ("[SUBMITTED] regulator: core: avoid unused variable warning")
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit bb41897e38c53458a88b271f2fbcd905ee1f9584)
Signed-off-by: Guenter Roeck <groeck@chromium.org>

Change-Id: Id6cf77332ee35bbb0bb56519c7aa595876b6963b
Reviewed-on: https://chromium-review.googlesource.com/331541
Reviewed-by: Filipe Brandenburger <filbranden@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>