UPSTREAM: mtd: spi-nor: wait until lock/unlock operations are ready
On Micron and Numonyx devices, the status register write command
(WRSR), raises a work-in-progress bit (WIP) on the status register.
The datasheets for these devices specify that while the status
register write is in progress, the status register WIP bit can still
be read to check the end of the operation.

This commit adds a wait_till_ready call on lock/unlock operations,
which is required for Micron and Numonyx but should be harmless for
others. This is needed to prevent applications from issuing erase or
program operations before the unlock operation is completed.

Reported-by: Stas Sergeev <stsp@list.ru>
Signed-off-by: Ezequiel Garcia <ezequiel@vanguardiasur.com.ar>
Signed-off-by: Brian Norris <computersforpeace@gmail.com>
(cherry picked from commit 32321e950d8a237d7e8f3a9b76220007dfa87544)
Signed-off-by: Wei-Ning Huang <wnhuang@google.com>

BUG=chrome-os-partner:40208
TEST=`emerge-oak chromeos-kernel-3_18`

Change-Id: If60a284ad36c96c7e9aaf2409d4e96b3742b8e1d
Reviewed-on: https://chromium-review.googlesource.com/330046
Commit-Ready: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>