UPSTREAM: drm/tegra: sor: Do not setup DPAUX at probe time
Setting up the DPAUX at probe time doesn't work because DPAUX is in the
same power domain as the SOR, and nothing may have turned on that domain
if the SOR driver didn't probe yet. The result is that register accesses
hang. Since the SOR device depends on the DPAUX device it's pretty much
guaranteed that nothing in the Linux kernel will have enabled the power
domain. The reason why this hasn't caused any problems is that the power
partition is enable by default on hardware reset.

However, firmware exists that will aggressively disable power partitions
before passing control to the kernel, at which point this becomes a real
problem. The solution implemented in this patch is to defer programming
of the mux option until the DPAUX pads are used. There is already an API
to do this, so just pass it a parameter that determines in which mode to
setup the pads. This will automatically make sure that no registers are
accessed before the SOR needs the DPAUX and has already enable the power
partition.

Fixes: 3227166c3bd6 ("drm/tegra: dpaux: Configure pads as I2C by
default")
Signed-off-by: Thierry Reding <treding@nvidia.com>
[from https://github.com/thierryreding/linux.git]
(cherry picked from commit 31377b1a08b5dd35841fabf6e6ef36034e7fefc6)

Change-Id: I722527260d711de41ad656f707eb71a1dc05dd99
Signed-off-by: Kary Jin <karyj@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/331022
Commit-Ready: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>