UPSTREAM: i2c: qup: Transfer each i2c_msg in i2c_msgs without a stop bit
The definition of i2c_msg says that

"If this is the last message in a group, it is followed by a STOP.
Otherwise it is followed by the next @i2c_msg transaction segment,
beginning with a (repeated) START"

So the expectation is that there is no 'STOP' bit inbetween individual
i2c_msg segments with repeated 'START'. Adding the support for the same.

This is required for some clients like touchscreen which keeps
incrementing counts across individual transfers and 'STOP' bit inbetween
resets the counter, which is not required.

This patch adds the support in non-dma mode.

Signed-off-by: Sricharan R <sricharan@codeaurora.org>
Reviewed-by: Andy Gross <andy.gross@linaro.org>
Tested-by: Archit Taneja <architt@codeaurora.org>
Tested-by: Telkar Nagender <ntelkar@codeaurora.org>
[wsa: updated commit message]
Signed-off-by: Wolfram Sang <wsa@the-dreams.de>
(cherry picked from commit f74187932d30e47533271554d79de7121cefb804
 git://git.kernel.org/pub/scm/linux/kernel/git/wsa/linux.git)

BUG=chrome-os-partner:49255
TEST=i2c eeprom works

Change-Id: Ice055ea2c465671dea05e5e76da56b2242170a59
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/333033
Commit-Ready: Matthew McClintock <mmcclint@qca.qualcomm.com>
Tested-by: Matthew McClintock <mmcclint@qca.qualcomm.com>
Reviewed-by: Grant Grundler <grundler@chromium.org>