UPSTREAM: mtd: spi-nor: add TB (Top/Bottom) protect support
Some flash support a bit in the status register that inverts protection
so that it applies to the bottom of the flash, not the top. This yields
additions to the protection range table, as noted in the comments.

Because this feature is not universal to all flash that support
lock/unlock, control it via a new flag.

Signed-off-by: Brian Norris <computersforpeace@gmail.com>
Tested-by: Ezequiel Garcia <ezequiel@vanguardiasur.com.ar>
(cherry picked from commit 3dd8012a8eeb3702fa17450ec1a16a3f38af138d)
Signed-off-by: Brian Norris <briannorris@chromium.org>

BUG=chrome-os-partner:40208
TEST=`emerge-oak chromeos-kernel-3_18`

Change-Id: I00813239238dec76d23727750933bb007a08f7c8
Reviewed-on: https://chromium-review.googlesource.com/330063
Commit-Ready: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>