BACKPORT: ASoC: rt5514: add rt5514 codec driver
This is the initial codec driver for rt5514. The codec includes a low power
DSP for voice wake up. The register address is incremental by 4 in the DSP
memory map. In order to recover the codec settings in the codec mode and
manipulate the DSP mode for voice wake up, we use the multi-level register
map. One is for ALSA API in codec mode that can be recovered by cache
before recording. Another is for DSP related settings that can be accessed
with 32bit address of the DSP in the application of voice wake up.

Signed-off-by: Oder Chiou <oder_chiou@realtek.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 4a6180ea7399b945cd380dc63e2e8118f9b432d3)

[pc.liao: req_sequence is not defined in Chrome OS 3.18, so use reg_default instead]

BUG=chrome-os-partner:50078
TEST=oak-rev6, sound capture works

Change-Id: I741bcd29452250c81d3aa6fd63e617ff657774d5
Signed-off-by: Koro Chen <koro.chen@mediatek.com>
Signed-off-by: PC Liao <pc.liao@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/321582
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>