UPSTREAM: tytso: ext4 crypto: move context consistency check to ext4_file_open()
In the case where the per-file key for the directory is cached, but
root does not have access to the key needed to derive the per-file key
for the files in the directory, we allow the lookup to succeed, so
that lstat(2) and unlink(2) can suceed.  However, if a program tries
to open the file, it will get an ENOKEY error.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
(cherry picked from commit 6df2762f3f76c580097add08fa03d97ed3bf57d7)
(From git tree: git://git.kernel.org/pub/scm/linux/kernel/git/tytso/ext4.git)

BUG=chromium:347322
TEST=Compiled, Installed on samus.

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: I561b78f71981e4b167c3832e44ef7d8c3dc9f7a7
Reviewed-on: https://chromium-review.googlesource.com/327218
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>