CHROMIUM: drm/nouveau: For iommu mapping, map size of mem
Rather than mapping the region described by the nvkm_vma, map the region
described by nvkm_mem.

In the case of partial mapping, we may have many nvkm_vma's that describe
subregions of the nvkm_mem.  However, we want to map the entire nvkm_mem
region once to the iommu.

When nvkm_vm_map_sg_table_with_iommu() is entered again with a different
subregion nvkm_vma, we will seemingly redo the iommu mapping for the same
nvkm_mem region...however, this iommu mapping is refcounted, so we won't do
the iommu mapping again redundantly.  In fact, we rely on this refcounting
so that when we eventually call mmu->unmap_iommu(), these unmaps get
ref-uncounted properly and the iommu mapping undone only once at the last
nvkm_vma unmapping.

In the more normal case of the nvkm_vma spanning exactly the region
described by the nvkm_mem, there is no change in behavior.

BUG=chrome-os-partner:49979
TEST=Build and boot to desktop on A44

Change-Id: Ib4dc3162b680f42035abdef640eff3b7e3ba877f
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/329878
Reviewed-by: Vince Hsu <vince.h@nvidia.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>