drm/nouveau: fifo: unify chan kick for gf100/gk104
Kicking (i.e., preempting) a channel is common for gf100 and gk104, so
move it to gf100_fifo_chan_kick, and make that public.

The logic for gf100_fifo_context_detach kick was slightly different
(wait for chid to disappear from the reg), but waiting for the pending
bit to go zero is a valid way for both (and more appropriate).

The pmu mutex locking is now also common, although not used for some
platforms.

Change-Id: I264bf388153b99789f514c08436abd64e2bb73b6
Signed-off-by: Konsta Holtta <kholtta@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/307240
Commit-Ready: Vince Hsu <vince.h@nvidia.com>
Reviewed-by: Vince Hsu <vince.h@nvidia.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>