drm/nouveau: Make the kthreads freezable
In case of system suspend, it is currently possible that pushbuffer
threads which have some pushbuffers in their queues will start accessing
the hardware before the device is resumed. To prevent this from
happening, this patch makes the threads freezable.

Runtime suspend is not affected, because all the code accessing the
hardware is surrounded by pm_runtime_get_sync() and
pm_runtime_put_autosuspend().

BUG=chrome-os-partner:48478
BUG=chrome-os-partner:49509
TEST=Runtime suspend/resume cycles running for >30 min did not lock the system up
TEST=Tried several system suspend/resume cycles and nothing crashed

Change-Id: I81c78d1a8ed843c35f63e4c6d1a4e59e0a3df156
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/329312
Reviewed-by: David Ung <davidu@nvidia.com>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>