UPSTREAM: i2c: qup: Fix fifo handling after adding V2 support
After the addition of V2 support, there was a regression observed
when testing it on MSM8996. The reason is driver puts the controller
in to RUN state and writes the data to be 'tx' ed in fifo. But controller
has to be put in to 'PAUSE' state and data has to written to fifo. Then
should be put in to 'RUN' state separately.

Signed-off-by: Sricharan R <sricharan@codeaurora.org>
Tested-by: Pramod Gurav <gpramod@codeaurora.org>
Signed-off-by: Wolfram Sang <wsa@the-dreams.de>
(cherry picked from commit 52db223e4cf414e216890643cda33e2626554471
 git://git.kernel.org/pub/scm/linux/kernel/git/wsa/linux.git)

BUG=chrome-os-partner:49255
TEST=i2c eeprom works

Change-Id: I7237490a71f0e679425d14db1fcd5c4c5a12a685
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/333035
Commit-Ready: Matthew McClintock <mmcclint@qca.qualcomm.com>
Tested-by: Matthew McClintock <mmcclint@qca.qualcomm.com>
Reviewed-by: Grant Grundler <grundler@chromium.org>