FROMLIST: ASoC: dapm: Make sure we have a card when displaying component widgets
The dummy component is reused for all cards so we special case and don't
bind it to any of them.  This means that code like that displaying the
component widgets that tries to look at the card will crash.  In the
future we will fix this by ensuring that the dummy component looks like
other components but that is invasive and so not suitable for a fix.
Instead add a special case check here.

Backdrop:
All components are initially given an empty card, and since commit 5897b9e118fa
("UPSTREAM: ASoC: core: Don't probe the component which is dummy")',
snd-soc-dummy will not be probed so that it remains an empty card assigned.

BUG=chrome-os-partner:49409
TEST=grep -rsI "hello" /sys/devices/platform/skl_nau88l25_ssm4567_i2s/

Reported-by: Harry Pan <harry.pan@intel.com>
Suggested-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Mark Brown <broonie@kernel.org>
(am from http://www.spinics.net/lists/alsa-devel/msg48042.html)
Signed-off-by: Harry Pan <harry.pan@intel.com>

Change-Id: Icaa6d5c8d39f087f85bf091fb6f27e71c3644de5
Reviewed-on: https://chromium-review.googlesource.com/331285
Commit-Ready: Harry Pan <harry.pan@intel.com>
Tested-by: Harry Pan <harry.pan@intel.com>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>