msm: emac: Add the support for external PP2S interrupt.
There is an open issue that no EMAC PPS_IN interrupt is asserted for
a long period of time. It impacts the PTP software on FSM9900 to do
the fine adjustment on EMAC RTC clock. To work around this issue, add
the support for external PP2S interrupt in the EMAC driver so that
the fine adjustment can be done when PP2S interrupt occurs.

Add the sysfs node to display the PTP interrupt count.

Change-Id: If07c03bee7663c406bd850713d43b82ae1057a13
Acked-by: Jie Luo <jluo@qti.qualcomm.com>
Signed-off-by: Sivanesan Rajapupathi <srajap@codeaurora.org>