FROMLIST: keyrings: Allow searching the user session keyring
Currently, if a session keyring exists, we are not searching in the
user session or user keyrings.

This is a problem when a session keyring exists and we want to use
ecryptfs.

BUG=chromium:347322,chromium:593893
TEST=Without this change, mounting an ecryptfs "partition" fails when a
session keyring exists:
kernel: [ 2686.047522] Could not find key with description: [dd6f92bd8660b36c]
Although the key exits:
keyctl show @us
Keyring
 549666721 --alswrv      0 65534  keyring: _uid_ses.0
 346719914 --alswrv      0 65534   \_ keyring: _uid.0
 235623693 --alswrv      0     0       \_ user: dd6f92bd8660b36c
 747773852 --alswrv      0     0       \_ user: 7025717e50fd74a2
With this change, ecryptfs can see the keys it needs.

Note that 'keyctl show' still only shows the session keyring by default.
We need to specify 'keyctl show @us' to see the user session keyring
when the session keyring exits.

[http://www.spinics.net/lists/keyrings/msg00647.html]

Change-Id: I8c6a112a77b7ba40f07e6a579cc77d88af44146a
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/333383
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>