UPSTREAM: memremap: don't modify flags
These patches implement a MEMREMAP_WC flag for memremap(), which can be
used to obtain writecombine mappings.  This is then used for setting up
dma_coherent_mem regions which use the DMA_MEMORY_MAP flag.

The motivation is to fix an alignment fault on arm64, and the suggestion
to implement MEMREMAP_WC for this case was made at [1].  That particular
issue is handled in patch 4, which makes sure that the appropriate memset
function is used when zeroing allocations mapped as IO memory.

This patch (of 4):

Don't modify the flags input argument to memremap(). MEMREMAP_WB is
already a special case so we can check for it directly instead of
clearing flag bits in each mapper.

Signed-off-by: Brian Starkey <brian.starkey@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
(cherry picked from commit 498a647fdb8f467aed8f52d89eb9b07a7b492380
 git://git.kernel.org/pub/scm/linux/kernel/git/mhocko/mm.git since-4.4)
Signed-off-by: Andrew-CT Chen <andrew-ct.chen@mediatek.com>

BUG=chrome-os-partner:49541
TEST=build oak, strago, smaug and allnoconfig, and boot oak-rev3

Change-Id: Ied971c9f850b34b914fd4d293c68099d83db7e43
Reviewed-on: https://chromium-review.googlesource.com/332410
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>