CHROMIUM: HACK: ASoC: Rename wake on voice device
The WoV implementation needs FW to be delivered.
So now if we enable OK GOOGLE, there is a fw crash.

cras looks for device named Wake on Voice.
Hence renaming the device until full solution is available
and will revert this once done.
This ensures DSP is not in error state.

BUG=chrome-os-partner:51052,chrome-os-partner:51020
TEST=enable OK GOOGLE and verify audio works across boot

Change-Id: I98a48eeb67eb85fe41fb20d93cd1fe0930cf073a
Signed-off-by: Sathya Prakash M R <sathya.prakash.m.r@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/332381
Commit-Ready: Ben Zhang <benzh@chromium.org>
Tested-by: Naveen M <naveen.m@intel.com>
Reviewed-by: Naveen M <naveen.m@intel.com>
Reviewed-by: Ben Zhang <benzh@chromium.org>