CHROMIUM: LSM: make modules FS pinning work with unshared mount namespace
Currently pinning modules filesystem is done by storing pointer to
"struct vfsmount" and comparing it to vfsmount the module is loaded from.
This does not work in presence of private mount namespaces.

This patch changes Chromium LSM so that instead of vfsmount it keeps a pointer
to super_block. Super block is unique for each mounted filesystem and stays
the same across private mount namespacess.

BUG=b:27720867
TEST=modules load from "unshared" mount namespace without errors

Signed-off-by: Andrey Ulanov <andreyu@google.com>
Reviewed-on: https://chromium-review.googlesource.com/334911
Reviewed-by: Kees Cook <keescook@chromium.org>
Change-Id: Iee60cb4ee1f3180cce0c88f54f4a50db9874e008
Reviewed-on: https://chromium-review.googlesource.com/341744
Reviewed-by: Olof Johansson <olofj@chromium.org>