UPSTREAM: mmc: Allow forward compatibility for eMMC
Reference CL 317473
As stated by the eMMC 5.0 specification, a chip should not be rejected
only because of the revision stated in the EXT_CSD_REV field of the
EXT_CSD register.
Remove the control on this value, the control of the CSD_STRUCTURE field
should be sufficient to reject future incompatible changes.

Signed-off-by: AdamTsai <adam.tsai@quantatw.com>

BUG=chrome-os-partner:48738
TEST=try a newish(rev 8) emmc, should enumerate just fine

Change-Id: I42ec8b5138b564d0c3dff8f71423e3ae327f73c0
Change-Id: Ibe596d20c520eeabf6340c3b99e2b8a868946ace
Reviewed-on: https://chromium-review.googlesource.com/336100
Reviewed-by: Bowgo Tsai <bowgotsai@google.com>
Reviewed-by: Chih-Wei Tsai <adam.tsai@quantatw.com>
Tested-by: Chih-Wei Tsai <adam.tsai@quantatw.com>
Commit-Queue: Chih-Wei Tsai <adam.tsai@quantatw.com>