Fixup: mfd: cros_ec_pd_update: Defer adding PD if needed
If cros_ec_pd_add() is called too early, pd_ec may not yet be initialized.
This results in a null pointer access and crash. Defer probing in this
situation.

As a side effect, the check if pd_ec is NULL in cros_ec_pd_update_check()
is no longer needed, since we never get there if pd_ec is NULL.

Signed-off-by: Guenter Roeck <groeck@chromium.org>