latent_entropy: Mark functions with __latent_entropy
The __latent_entropy gcc attribute can be only on functions and variables.
If it is on a function then the plugin will instrument it. If the
attribute is on a variable then the plugin will initialize it with a
random value.  The variable must be an integer, an integer array type or a
structure with integer fields.

These functions have been selected because they are init functions, are
called at random times, or they have variable loops, each of which provide
some level of latent entropy.

Signed-off-by: Emese Revfy <re.emese@gmail.com>
Signed-off-by: Kees Cook <keescook@chromium.org>