NFS: Getattr doesn't require data sync semantics
When retrieving stat() information, NFS unfortunately does require us to
sync writes to disk in order to ensure that mtime and ctime are up to
date. However we shouldn't have to ensure that those writes are persisted.

Relaxing that requirement does mean that we may see an mtime/ctime change
if the server reboots and forces us to replay all writes.

The exception to this rule are pNFS clients that are required to send
layoutcommit, so we add a helper nfs_filemap_write_and_wait() that can
call pnfs_sync_inode() after the pagecache has been synced.

Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>