CHROMIUM: input: cros_ec_keyb: Add non-matrix buttons and switches
On some newer boards using mkbp we're hooking up non-matrix buttons and
switches to the EC but NOT to the main application processor.  The EC
has recently had some new commands added for allowing the kernel to work
with these buttons and switches, namely:
- http://crosreview.com/358633
- http://crosreview.com/358634
- http://crosreview.com/358926

Let's add kernel support to handle this.  Rather than creating a whole
new input driver, we'll continue to use cros_ec_keyb and just report the
new keys.

BUG=chrome-os-partner:54976, chrome-os-partner:53509
TEST=Try lid, power, and volume on kevin/gru w/ evtest
TEST=Boot lid closed: evtest --query /dev/input/event1 EV_SW SW_LID; echo $?
TEST=Suspend/resume w/ lid while running evtest
TEST=Try on veyron; no trouble found
TEST=check Phys= entries in `cat /proc/bus/input/devices`

Change-Id: Id57f1e80ee9772b0fc5fc3fdd2a5993e9a45f31b
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/359059
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@google.com>

Conflicts:
	drivers/input/keyboard/cros_ec_keyb.c

[rebase49(groeck): Resolved conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>